move CommonVectorStoreProperties to .vectorstore. package
