+++ spring-ai-spring-boot-autoconfigure/src/main/java/org/springframework/ai/autoconfigure/vectorstore/CommonVectorStoreProperties.java
@@ -1,4 +1,4 @@
-package org.springframework.ai.autoconfigure;
+package org.springframework.ai.autoconfigure.vectorstore;
 
 /**
  * @author Josh Long
